Move the linkset normalization types out of the internal dir.
